requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
langgraph
langchain
python-dotenv
pymysql
langchain_openai
